[promotionAssistant]搜索词报告功能和其他功能bug修复

+++ perfect-service/src/main/java/com/perfect/utils/RegionalCodeUtils.java
@@ -2,11 +2,9 @@ package com.perfect.utils;
 
 import com.google.gson.Gson;
 import com.google.gson.reflect.TypeToken;
-import com.perfect.entity.StructureReportEntity;
 import com.perfect.redis.JRedisUtils;
 import redis.clients.jedis.Jedis;
 
-import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -16,25 +14,25 @@ import java.util.Map;
  */
 public class RegionalCodeUtils {
 
-    public static  Map<Long,String> regionalCode(List<Long> id){
+    public static  Map<Integer,String> regionalCode(List<Integer> id){
         Jedis jc = JRedisUtils.get();
         String data = jc.get("RegionalCodeRedis");
         Gson gson = new Gson();
-        Map<String,String> d = gson.fromJson(data, new TypeToken<Map<Long,String>>() {}.getType());
-        Map<Long,String> map = new HashMap<>();
-        for (Long aLong : id){
+        Map<String,String> d = gson.fromJson(data, new TypeToken<Map<Integer,String>>() {}.getType());
+        Map<Integer,String> map = new HashMap<>();
+        for (Integer aLong : id){
             map.put(aLong, d.get(aLong));
         }
         return map;
     }
-    public static  Map<Long,String> regionalCodeName(List<String> name){
+    public static  Map<Integer,String> regionalCodeName(List<String> name){
         Jedis jc = JRedisUtils.get();
         String data = jc.get("RegionalCodeRedis");
         Gson gson = new Gson();
-        Map<Long,String> d = gson.fromJson(data, new TypeToken<Map<Long,String>>() {}.getType());
-        Map<Long,String> map = new HashMap<>();
+        Map<Integer,String> d = gson.fromJson(data, new TypeToken<Map<Integer,String>>() {}.getType());
+        Map<Integer,String> map = new HashMap<>();
         for (String aString : name){
-            for (Map.Entry<Long, String> voEntity : d.entrySet()){
+            for (Map.Entry<Integer, String> voEntity : d.entrySet()){
                 if(aString.equals(voEntity.getValue())){
                     map.put(voEntity.getKey(),aString);
                 }
